requirements.txt
pyowm
telebot
datetime
python-dotenv
python-telegram-bot
schedule
